requirements.txt
# OmniBox - Dependencies
# Install with: pip install -r requirements.txt

# Core AI
google-generativeai>=0.3.0

# Environment & Config
python-dotenv>=1.0.0

# HTTP Requests
requests>=2.28.0

# Speech (Text-to-Speech & Speech-to-Text)
pyttsx3>=2.90
SpeechRecognition>=3.10.0
pyaudio>=0.2.13

# Screenshot
pyautogui>=0.9.53

# Terminal Colors
colorama>=0.4.6